# backend/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API REST para gestão de empresas, pesquisas, dashboards, auditoria e submissões anônimas.\n\nFluxo sugerido para clientes frontend:\n1) Autenticar em /auth/login\n2) Enviar token JWT no header Authorization: Bearer \u003ctoken\u003e\n3) Consumir endpoints protegidos em /api/v1\n",
        "title": "API de Pesquisa de Clima Organizacional",
        "termsOfService": "https://example.com/terms",
        "contact": {
            "name": "Equipe Backend",
            "email": "backend@example.com"
        },
        "license": {
            "name": "Proprietary"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/v1/dashboards": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dashboards"
                ],
                "summary": "Criar dashboard",
                "parameters": [
                    {
                        "description": "Dados do dashboard",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/organizational-climate-survey_backend_internal_application_dto.DashboardCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/organizational-climate-survey_backend_internal_application_dto_response.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/organizational-climate-survey_backend_internal_application_dto_response.APIResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/organizational-climate-survey_backend_internal_application_dto_response.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/organizational-climate-survey_backend_internal_application_dto_response.APIResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/dashboards/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dashboards"
                ],
                "summary": "Buscar dashboard por ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID do dashboard",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/organizational-climate-survey_backend_internal_application_dto_response.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/organizational-climate-survey_backend_internal_application_dto_response.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/organizational-climate-survey_backend_internal_application_dto_response.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/organizational-climate-survey_backend_internal_application_dto_response.APIResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dashboards"
                ],
                "summary": "Atualizar dashboard",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID do dashboard",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Dados para atualização",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/organizational-climate-survey_backend_internal_application_dto.DashboardUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/organizational-climate-survey_backend_internal_application_dto_response.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/organizational-climate-survey_backend_internal_application_dto_response.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/organizational-climate-survey_backend_internal_application_dto_response.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/organizational-climate-survey_backend_internal_application_dto_response.APIResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dashboards"
                ],
                "summary": "Remover dashboard",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID do dashboard",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/organizational-climate-survey_backend_internal_application_dto_response.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/organizational-climate-survey_backend_internal_application_dto_response.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/organizational-climate-survey_backend_internal_application_dto_response.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/organizational-climate-survey_backend_internal_application_dto_response.APIResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/dashboards/{id}/refresh": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dashboards"
                ],
                "summary": "Atualizar dados do dashboard",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID do dashboard",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/organizational-climate-survey_backend_internal_application_dto_response.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/organizational-climate-survey_backend_internal_application_dto_response.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/organizational-climate-survey_backend_internal_application_dto_response.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/organizational-climate-survey_backend_internal_application_dto_response.APIResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "organizational-climate-survey_backend_internal_application_dto.DashboardCreateRequest": {
            "type": "object",
            "required": [
                "id_pesquisa",
                "titulo"
            ],
            "properties": {
                "config_filtros": {
                    "description": "Configuração JSON de filtros (opcional)",
                    "type": "string",
                    "example": "{\"setor\":\"RH\"}"
                },
                "id_pesquisa": {
                    "description": "Identificador da pesquisa associada (obrigatório)",
                    "type": "integer",
                    "example": 10
                },
                "titulo": {
                    "description": "Título do dashboard (obrigatório)",
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 3,
                    "example": "Dashboard Clima RH"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}